[시작페이지] 코드 관심사의 분리 (#338)
* refactor: scroll 애니메이션 코드 분리 및 애니메이션 폴더 생성

* refactor: 첫번째 선물 애니메이션 코드 분리 및 start style 페이지 정리

* refactor: footer position Footer의 고정값으로 수정

* refactor: 시작페이지 고정 버튼 코드분리

* style: start button 주석 삭제

* style: start button 컴포넌트 명 변경

* fix: 글로벌 스타일에서 margin 0 없앰

---------
