+++ src/pages/Start/StartAnimation/ScrollAnimation/StartScrollAnimation.style.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const PersonWrapper = styled.div`
+  position: absolute;
+  top: 86rem;
+  z-index: 1;
+`;
