+++ src/style/GlobalStyle.ts
@@ -29,7 +29,8 @@ body {
 #root, body, html {
     scrollbar-width: none; /* 파이어폭스 스크롤바 숨김 */
     
-    margin: 0 auto;
+    height: 100vh; 
+    width: 100vw;
     padding:0;
     font-size: 62.5%;
     -ms-overflow-style: none; /* 인터넷 익스플로러  스크롤바 숨김 */
